modified UserController, routes,reply.js, created a new table report_feedbacks.php and models. Change name of FeedbackSpamCount to FeedbackReply

+++ app/controllers/UserController.php
@@ -3,7 +3,8 @@
 class UserController extends BaseController {
 
 	public function __construct(User $user, Subscribe $subscribes, Notification $notification, 
-		Video $video, UserWatchLater $watchLater, UserFavorite $favorite, Feedback $feedback,Playlist $playlist){
+		Video $video, UserWatchLater $watchLater, UserFavorite $favorite, Feedback $feedback,Playlist $playlist,
+		ReportedFeedback $reportedFeedback){
 		$this->Notification = $notification;
 		$this->Video = $video;
 		$this->Subscribe = $subscribes;
@@ -14,6 +15,7 @@ class UserController extends BaseController {
 		$this->Favorite = $favorite;
 		$this->Feedback = $feedback;
 		$this->Playlist = $playlist;
+		$this->ReportedFeedback = $reportedFeedback;
 	}	
 
 	public function getSignIn() { 
@@ -633,15 +635,13 @@ class UserController extends BaseController {
 	public function getViewUsersFeedbacks($channel_name) {
 		$user_id = 0;
 		$userChannel = User::where('channel_name', $channel_name)->first();
-		
-
 		$userFeedbacks = $this->Feedback->getFeedbacks($userChannel->id);
 
 		foreach ($userFeedbacks as $key => $userFeedback) {
 			$userFeedbacks[$key]->img = $this->User->addProfilePicture($userFeedback->user_id);
 			$userFeedbacks[$key]->likesCount = DB::table('feedbacks_likesdislikes')->where(array('feedback_id' => $userFeedback->id, 'status' => 'liked'))->count();
 			$userFeedbacks[$key]->dislikeCount = DB::table('feedbacks_likesdislikes')->where(array('feedback_id' => $userFeedback->id, 'status' => 'disliked'))->count();
-
+			if(Auth::check()){
 			$userFeedbacks[$key]->ifAlreadyLiked = DB::table('feedbacks_likesdislikes')->where(array(
 				'feedback_id' => $userFeedback->id, 
 				'user_id' => Auth::User()->id,
@@ -652,16 +652,21 @@ class UserController extends BaseController {
 				'user_id' => Auth::User()->id,
 				'status' => 'disliked'
 				))->first();
-
-			$userFeedbacks[$key]->getFeedbackReplies = DB::table('feedbacks_replies')
-			->join('users', 'users.id', '=', 'feedbacks_replies.user_id')
+			}
+			$userFeedbacks[$key]->countFeedbackReplies = DB::table('feedback_replies')
+			->join('users', 'users.id', '=', 'feedback_replies.user_id')
 			->where('feedback_id', $userFeedback->id)->count();
+
+			$userFeedbacks[$key]->getFeedbackReplies = FeedbackReply::select('feedback_replies.id', 'user_id', 'reply', 'feedback_id',
+			 'feedback_replies.created_at', 'feedback_replies.updated_at', 'channel_name')
+			->join('users', 'users.id', '=', 'feedback_replies.user_id')
+			->get();	
 			
-			// foreach($replies->get() as $key => $reply){
-			// 	$replies->get()[$key]->img = $this->User->addProfilePicture($reply->user_id);
-			// }
-		
+			$userFeedbacks[$key]->spamCounts = DB::table('reported_feedbacks')->where('feedback_id', $userFeedback->id)->count();
 		}
+
+		// return $userFeedbacks;
+
 
 		$allViews = DB::table('videos')->where('user_id', $userChannel->id)->sum('views');
 		$countAllViews = $this->Video->countViews($allViews);
@@ -839,19 +844,39 @@ public function getDeleteFeedback() {
 	$channelId = Input::get('channel_id');
 	$userId = Input::get('user_id');
 	$feedback_id = Input::get('feedback_id');
+	$id = Input::get('id');
+	$deleteFeedback = DB::table('feedbacks')->where(array('channel_id' => $channelId, 'user_id' => $userId, 'id' => $feedback_id))->delete();
+	return Response::json($deleteFeedback);
+}
 
+public function postDeleteFeedbackReply() {
+	$feedbackId = Input::get('feedback_id');
+	$userId = Input::get('user_id');
+	$id = Input::get('id');
 
-	$deleteFeedback = DB::table('feedbacks')->where('id',$feedback_id)->delete();
+	$a = FeedbackReply::find($id)->delete();
+	return Response::json($a);
 
-	return Response::json(array('status' => 'sucess', 'id' => $feedback_id));
 }
 
 public function postSpamFeedback() {
-	return Input::all();
+
 	$channelId = Input::get('channel_id');
 	$userId = Input::get('user_id');
-	$feedbackId = Input::get('feedbackId');
-	$a = Feedback::where('id',$feedbackId)->first();
+	$spamId = Input::get('spamID');
+	$a = $this->ReportedFeedback->getReportCount($spamId, $channelId, $userId);
+
+	
+	return Response::json($a);
+}
+
+public function postSpamFeedbackReply() {
+
+	$id = Input::get('reportID');
+	$user_id = Input::get('user_id');
+	$created_at = date('Y-m-d H:i:s');
+	$updated_at = date('Y-m-d H:i:s');
+	$a = DB::table('reported_replies')->insert(array('reply_id' => $id, 'user_id' => $user_id, 'created_at' => $created_at, 'updated_at' => $updated_at));
 
 	return Response::json($a);
 }
